TIKA-4154 -- parameterize json max string length -- fix unit test in tika-server-standard (#1403)
 TIKA-4154 -- update unit test in tika-server-standard

(cherry picked from commit e1f207a4aea434d57385e788dcd5f882f9206248)
